got 'vt-dev-load-theme-frame-2' command working.  This will set the theme for all frames that have been set with 'vt-dev-load-theme-frame-2' (via shared hash vt-dev-custom-frame-theme-hash).  This is kind of a hack to allow you setting emacs 24 themes at the frame level.  Ideally, you would like to just set it for one frame, but I kept getting side effect issues where the other frames would sometimes change colors when you attempted to just alter one.
To get around this, I just hacked so that all frames are set with
their specified frame everytime *any* frame theme is altered.

The code is currently ugly but works.  I want to check in to a
common repository so I can propagate to some of my other machines.

The code needs a serious cleanup.